fix(GO): fix type mismatch for nested args
In a previous PR the assumed type of nested arg was a map with string keys and interface typed values which raised mismatch error. This PR fixes it by asserting keys to be of type interface as well.

fix(GO): fix type mismatch error

fix(CLI): retrigger release for GO